requirements.txt
opencv-python==4.8.0.76
face-recognition==1.3.0
mysql-connector-python==8.0.33
